feat(socialStats): #56 Fixed async userSocial doc creation issue
