+++ web/src/app/app.component.ts
@@ -1,8 +1,8 @@
-import { Component, OnInit, NgZone } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { environment } from '../environments/environment';
-import { AngularFireAuth } from '@angular/fire/auth';
-
+import { UserService } from './services/user.service';
+import { UserSocial } from './models/userSocial.model';
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
@@ -14,19 +14,53 @@ export class AppComponent implements OnInit {
 
   constructor(
     private router: Router,
-    private afAuth: AngularFireAuth,
-    private zone: NgZone,
+    private userService: UserService,
   ) {
     this.version = environment.version;
   }
 
   ngOnInit() {
-    this.afAuth.auth.onAuthStateChanged((user) => {
-      if (user) {
-        this.zone.run(() => {
-          this.router.navigate(['/dashboard']);
-        });
-      }
-    });
+    this.userService.getUser()
+      .subscribe((user) => {
+        if (user) {
+          const userSocial = this.userService.addRefID(user);
+          this.saveUserSocialDetails(userSocial, user.credentials.providerId.split('.')[0]);
+        }
+      });
+  }
+
+  saveUserSocialDetails(userData, provider: string) {
+    const socialDetails = {};
+    socialDetails[provider] = userData;
+    this.userService.checkForSocialDoc(provider, userData.userId)
+      .subscribe((response) => {
+        if (!response) {
+          console.error('error in checking of existing social doc');
+          return;
+        }
+        if (response.empty) {
+          this.createUserSocialDetails(socialDetails);
+          return;
+        }
+        this.updateUserSocialDetails(response.id, socialDetails);
+      });
+  }
+
+  createUserSocialDetails(socialDetails: UserSocial) {
+    this.userService.addSocialDoc(socialDetails)
+      .subscribe((response) => {
+        if (!response) {
+          console.error('error in creating social doc');
+          return;
+        }
+        this.router.navigate(['/dashboard']);
+      });
+  }
+
+  updateUserSocialDetails(id: string, socialDetails: UserSocial) {
+    this.userService.updateSocialDoc(id, socialDetails)
+      .subscribe(() => {
+        this.router.navigate(['/dashboard']);
+      });
   }
 }
