+++ web/src/environments/environment.ts
@@ -12,7 +12,7 @@ export const environment = {
     databaseURL: '',
     projectId: '',
     storageBucket: '',
-    messagingSenderId: ''
+    messagingSenderId: '',
   }
 };
